[DEV-657] session/base: add get_current_working_schema_version (#363)
Adds a basic get_current_working_schema_version for us to keep track of 
what the local working schema version is. This will eventually be compared 
with a version that we persist in the end-users working schema. We will then 
compare the two values to determine if we want to run the more expensive 
initialization functions.