session/base: rm filter in register functions and procedures (#382)
* session/base: rm filter in register functions and procedures

We recently added some logic to optimize the initialization flow
that is gated by a version number. As such, we don't need to optimize
this check anymore here anymore. This also allows us to update
stored functions and procedures when we update them, as we would
previously only write these functions if they were net new.

* test_snowflake_session: update unit tests

We can delete one of these tests since it's identical with
another one after we've removed the extra logic.

* session/base: add docstring