service/test_feature: autouse mock execute query (#388)
We were seeing some extra errors when trying to merge some of
the featuer store ID validation changes. This is because the
session object is shared between unit tests, and making any
updates to that was causing some errors since the updates were now
returning a new session for each test (functionally scoped).

If the session is functionally scoped, running individual unit tests
breaks since they're not explicitly mocked. As such, each unit tests
try to make a call, which is not what the test was expecting, since
the unit test was expecting a mocked call.

As such, we auto inject a mocked execute query fixture around
all the tests in this module to handle this automatically for us.