routes/base: add extra assertion for clarity
This test would occasionally fail, depending on how the column_info
fixture is updated. This commit adds some extra assertions to make
sure that the test fails with a clearer indication as to what
is expected from the fixture (i.e. the item_id column is the last
element in the array).