service/base_service: use DocumentServiceT
There was some duplication in types being used when trying to
create new data types. This felt a little tideous, and error prone.
Since the docstring also mentions that we want to return a
DocumentServiceT, I've updated the `as_service` function to return
the type, and also updated the type bound of the generic type
to be bound to BaseDocumentService.