[DEV-657] implement feature store ID validation (#409)
This PR adds some validation for matching feature store IDs, that is performed 
when a user tries to retrieve a session. This ensures that the feature store used for 
session they're retrieving, matches the feature store that is persisted in the working schema.

Note that this PR does NOT include the final changes for the validation and writing of 
the feature store ID to the working schema table when a user creates a FeatureStore via 
FeatureStore.get. I was facing some test issues when updating the code, and will resolve 
those issues while this PR is being reviewed (which is already a pretty big PR).

Details on the overall validation strategy can be found on this confluence doc - https://featurebyte.atlassian.net/wiki/spaces/DEV/pages/40632321/Session