session_validator: explicitly cast to dict
There looks to be a small difference with how our test mongo
client, and the production one interpret the details object in
a find lookup. This updates to explicitly cast to a dictionary
so that we can resolve an issue that is causing the functional
tests to fail when we bump the version on the featurebyte-app.