featurebyte/core: mv append_to_lineage to util
There didn't really seem to be a need for this function to be in
a mixin since it didn't have any dependencies on `self`, and looked
like it could be a pure util function. This also makes it easier
to reuse in other utils that I'm planning to have for adding the
join functionality into Views, since I don't need to add the mixin
just to be able to use this one function.