service/info: mv dependency on dataservice into class
This commit is a simple proof-of-concept to move info_service's
dependency on data_service, into the info_service class.

There are a few benefits
- we reduce a monolithic base service class.
- we move direct dependencies closer to the calling class,
  making it easier for a reader to see what other services, a
  particular service might depend on.
- we initialize inside the constructor, instead of a property, so
  that we don't have to repeatedly initialize the dependent service

More specifically, we can use the reduction in a monolithic base
service class for potential test optimizations in the future.

For example, a typical optimization would be to only run tests
that are affected by your code change. This is typically done by looking
at where a module is used, and running all tests that are associated
with that module.

Currently, if we make a change to the DataService, all
non-document services will require their tests to be re-run since they
all transitively depend on DataService through the BaseService.
This happens even though only the InfoService actually has any
dependency on the DataService.

More immediately as well, this also helps to reduce boilerplate
when adding new services. The main trade-off is that we don't
have all the document services easily accessible in a single place,
but that feels like an ok trade-off, especially since they
are already discoverable by looking at implementations of the
BaseDocumentService.