feature store validation: add ability to persist credentials (#477)
In order to be able to support feature store validation with the SaaS side of the platform, we are updating the FeatureStore#create endpoint, and the underlying API route, to take in an Optional[Credential] that users can specify when creating a feature store.

For the SDK side, this is entirely optional, as we will default to the config file to read the credentials. However, if

there's no credential entry in the config file, AND
a user enters their credentials via the SDK,
we will persist the credentials into the config file for them when we create the feature store.
For the SaaS side (after this PR is merged, and the SDK version is bumped in featurebyte-app), we'll make an update to persist the credentials to the credentials collection when they create a feature store. We will also update the UI to pass down the credentials when we create the feature store.

This PR makes the SDK side updates that are needed.