api/view: drop join key from right columns (#497)
The current resulting view contains the the join key column from both the calling, 
and the target views. This results in duplicated information that isn't really necessary.

As such, we update the logic here to always remove the join key from the target view, 
and will only retain the column from the calling view.

We have already updated the validation logic in a previous PR so we don't need to make 
any changes to that here.