featurebyte/app: mv dependency construction out of for loop
We repeatedly recreate all the dependencies. This doesn't look
like it's fully necessary since we already have a global dep
on the mongo connection, and our connections to the data warehouse's
are session scoped. As such, I don't think we need to reconstruct
instances of all the other services and controllers for each router
(unless there's a requirement on the FastAPI side).

This should help to make the app initialization more efficient.