[DEV-725] Move window aggregation and item aggregation logic out of `feature_compute.py` (#529)
This refactors feature_compute.py to be more general so that it is easier to
add new types of aggregations (next to be added: lookup features). Most of the
changes are moving files around.

Also updated to quote point-in-time column everywhere in the generated sql.
This is not strictly necessary but is more consistent and allows customisation
of point-in-time column in the future if needed.