app_container: inject ConfigCredentialProvider, and refactor app_container initialization (#553)
MVP for a refactor to some of app container initialization, and the get_credential logic.

This was a pain point I encountered when trying to update some of the feature store 
validation logic.

Benefits of encapsulating behind the credential provider class

longer term - we can migrate featurebyte-app to something similar. This way, we don't have 
to pass get_credential around too much, and it'll be easier for featurebyte-app to inject in its 
own implementation of the credential provider.

We can also avoid having to pass the Any type around for get_credential

The refactor to the app_container splits up the logic a little bit to make the dependency 
injection more extensible for other objects that aren't just services, or controllers.