[DEV-915] Derive row index lineage from graph automatically (#566)
This refactors the handling of row index lineage so that it is derived from the
query graph automatically.

After the changes, row index lineage is tracked as part of OperationStructure,
and each node implementation is responsible of determining how it changes the
row index lineage (most nodes don't except for Filter, Join, etc). There is no
need for manual book keeping of row index lineage when working with higher
level objects such as View or ViewColumn.