[DEV-910] Implement latest value aggregation with bounded windows (#559)
This implements the latest value aggregation with bounded windows. One example
of such features is "Latest transaction amount in the past 30 days".

Each aggregation method is tagged with whether it is order dependent. For order
dependent aggregation methods, the aggregation will be performed using window
functions with a PARTITION BY clause instead of the usual GROUP BY clause. This
handling is applied both when generating sql for building tiles and building
features.