[DEV-952] Fix invalid argument types for function 'OBJECT_AGG' error (#602)
This fixes an error that arises when creating features with latest aggregation
method per category, and a value column that is not numeric.

According to the documentation of OBJECT_AGG, the value column must be of
VARIANT type, though it seems it is not strictly enforced. This PR updates the
call the OBJECT_AGG to always cast the value column to VARIANT type.