[DEV-979] Preview column without using nested query in generated sql (#626)
Currently when previewing a column, the generated sql performs a nested query.
This approach doesn't always work because sometimes the expression to be
previewed is only valid within the same select statement (e.g. when the
expression comes from a joined table). This PR fixes that by avoiding the
nested query.