[DEV-1005] update validation on point in time (#692)
Previously, we would throw an error for simple aggregations when on `POINT_IN_TIME` is provided when trying to preview. This isn't needed as simple aggregations are not time based.

This PR updates the logic such that we only check for `POINT_IN_TIME` being passed in if the features being previewed are time based.

If the features are not time based, and no `POINT_IN_TIME` parameter is passed in, we default the point in time to an arbitrary time in the past.