[DEV-1065] Fix feature jobs being overwritten due to tile_id collision (#704)
This fixes the issue where feature jobs are overwritten because tile_id (tile
table name) is used as the unique identifier for scheduled tasks, but a tile
table is expected to be updated by different jobs corresponding to different
feature transformations. The main change is to schedule jobs using
aggregation_id as the unique identifier.