[DEV-1080] Implement sql generation for serving parent features (#739)
This implements sql generation for serving parent features using children
entities. When any of the required entities is not provided in feature
requests, we will attempt to use existing relationships to lookup the values of
the missing entities. This way, we can support materialising a feature using
its children entities.