[DEV-1131] Fix invalid identifier error related to add_feature on complex views (#797)
This fixes the invalid identifier error that can happen after calling
add_feature on an EventView that has been updated by joins. Currently, we have
an optimisation logic to select input nodes for single column project nodes in
order to avoid redundant nodes. However, that violates the assumptions made in
SQL generation logic for views involving joins.

That optimisation logic is not strictly required now that our pruning algorithm
has been improved compared to the initial version, so this PR removes it to fix
the issue.