[DEV-1141][DEV-1142] add RelationshipInfo to keep track of relationship metadata (#815)
This commit adds a new collection to keep track of specific relationships between entities. If all 
the parent-child relationships were a graph, this collection would represent the individual edges in 
it. This is to allow us to be able to surface relationships, and assorted metadata more easily. This 
will also allow us to enable and disable individual relationships.

See https://featurebyte.atlassian.net/wiki/spaces/DEV/pages/17072129/Child+Parent+Entity+Relationship 
for more details.

Note that this PR only adds the basic CRUD operations for dealing with the collection. We won't actually 
be writing to this until a later point. 

Further work to be done
- Integrate this into the existing write paths of creating relationships
- Write a migration to add links for existing relationships

See https://featurebyte.atlassian.net/browse/DEV-1140 for more details.