integration/conftest: yield and close event loop
This commit updates the event_loop fixture to explicitly call
loop.close(). This is to properly clean up any issues in the
event loop when running async tests.

I'm not entirely sure if this changes anything, but I did notice
some of the integration test flakes looked to be around the error
"There is no current event loop in thread".

As such, I went to explore this error message and noticed that
most guides suggest that we yield and close the event loop to
prevent unexpected behaviour. Hopefully we see less flakes on
the main branch.