Re-direct spark docker logs to a file (#867)
* Re-direct spark docker logs to a file

And use the upload-artifact action to save the log file on each
workflow run. I set the retention to 5 days since these logs can
get large and should really on be valuable when investigating
failures.

* Add back exit code handling

Since we are running with +e to have a chance to gran the logs after
a failure we need to make sure we preserve the exit code and exit
properly on falure.