[DEV-1100] add production_ready_validator (#895)
This PR adds back the main body of code for doing PRODUCTION_READY guardrails, but 
doesn't add it into the main update_feature code path for now. This is so that the bug fixes 
don't have to be reviewed together with the original src code, and can be made as smaller increments.

This PR should be a no-op on the existing workflows.