api/catalog: add create_entity to catalog
We want to add a create_entity function to the catalog.
We do this by first adding an `Entity.create` function, which
constructs an Entity object, and saves it in the same function
call (compared to the current state where users have to construct
their own entity object, then save it in a separate method call).

We then add the `create_entity` endpoint into the Catalog so that
it is more easily discoverable by users.