[DEV-1169] Fix add_feature ignoring fill_value specified when creating feature (#928)
In the current implementation of add_feature(), post-processing steps such as
fillna are applied before the feature is joined to the EventView. One
undesirable consequence is that fill_value (or any user specified
post-processing) is ignored for features added to EventView.

To resolve this issue, this reworks JoinFeature node's sql implementation so
that the behaviour of adding a feature to EventView is as if the EventView is
used as the request table while serving the non-time based feature.