fbautodoc/extension: fix return type
Return type values are sometimes missing from the signature.
This is because the type checker doesn't find the type as we do
some special type checking handling in the src code to workaround
circular imports.

This commit updates the logic to fallback to the type found in the
docstring instead if we cannot infer the type from the signature.