[RELEASE] 0.1.0.dev724 with 0.0.5 of beta (#962)
* api/catalog: rm *_namespace methods (#942)

These are duplicated with the other feature/feature_list methods.
As such, we can remove them.

* fbautodoc/extension: sync changes from docs repo (#944)

* [BETA] fixing user without ~/.docker/config.json (#943)

* fixing user without ~/.docker/config.json

* create ~/.docker folder if does not exist

* [DEV-1250] Display progress updates for async tasks in the SDK (#939)

* display progress updates for async tasks in the SDK

* import redis uri from utils

* [DEV-1263] send telemetry to https://log.int.featurebyte.com (#945)

* [DEV-1257] documentation: fix naming and improve css (#947)

This PR fixes some stuttering in the API path that appears in the docs.

Also removes a css override to give the docs a cleaner look.

* Revert "[DEV-1208] Use tablesample for sampling (#930)" (#946)

This reverts commit 1676134feb7756b4e2fcddd7351ca667173aa2b4.

* Revert "[DEV-1263] send telemetry to https://log.int.featurebyte.com (#945)" (#953)

This reverts commit 774953a85aade933db54da395b78123e19f9825e.

* [DEV-1274] Add primary_entity attribute to Feature, FeatureList and FeatureGroup (#949)

This adds the primary_entity attribute to Feature, FeatureList and FeatureGroup
objects. Without this, the primary entity information is only available to
saved objects via their info() methods.

* Split up the groupby class (#940)

* new branch stuff

* slowchanging changed to scd

* groupby_test changed

---------

Co-authored-by: rkodati <rkodati@umich.edu>

* [DEV-1264] Fix errors when creating additional lag columns in ChangeView (#955)

This fixes two issues when creating additional lag columns in ChangeView:

1. Lag of past value column is not blocked, but lagging a column more than once
is not supported

2. MISSING_COLUMN error in Spark when creating additional lag columns in
ChangeView

* fix datasets import SQL (#956)

* [DEV-1258] fbautodoc/extension: fix return type (#957)

Return type values are sometimes missing from the signature.
This is because the type checker doesn't find the type as we do
some special type checking handling in the src code to workaround
circular imports.

This commit updates the logic to fallback to the type found in the
docstring instead if we cannot infer the type from the signature.

* api/catalog: have list_feature and list_feature_list delegate correctly (#958)

We were delegating to the wrong function by accident. This commit
fixes that.

* [DEV-1272] api_object: update docs (#948)

Update docstrings for get, get_by_id, info and save.

* [DEV-1282] catalog: add get_data_source and get_view (#960)

Add a couple more convenience methods to make reduce boilerplate
in the SDK.

* [DEV-1276] Documentation container in featurebyte docker compose (#951)

* added documentation container and replace image tag in workflow

* add check that documentation image has already been published

* linting fixes

* moving variable declaration forward

* adding comments

* remove dependency from featurebyte-docs to featurebyte-server

* added docs_enabled flag to be removed after beta

---------

Co-authored-by: Jevon Yeoh <jevon@featurebyte.com>
Co-authored-by: Kenny Chua <kenny@featurebyte.com>
Co-authored-by: Yung Siang Liau <yungsiang@featurebyte.com>
Co-authored-by: Rohan Kodati <56242701+kodatiro@users.noreply.github.com>
Co-authored-by: rkodati <rkodati@umich.edu>