gen_ref_pages: simplify accessor logic for reuse
This commit simplifies the accessor logic to make it easier to reuse,
and also fixes some bugs.

This commit will now generate a markdown page for each class
that inherit the accessor. This will allow us to display the
API path properly, and also not have some link issues that was
previously occurring in the docs.