[DEV-1201] gen_ref_pages: simplify accessor logic for reuse (#965)
* gen_ref_pages: simplify accessor logic for reuse

This commit simplifies the accessor logic to make it easier to reuse,
and also fixes some bugs.

This commit will now generate a markdown page for each class
that inherit the accessor. This will allow us to display the
API path properly, and also not have some link issues that was
previously occurring in the docs.

* gen_ref_pages: rm all_markdown_files debug

This isn't needed anymore, and was also causing slow build times.
Clean up some of this code.

* Fix docstring