[DEV-1292] Fix error in on demand tile generation when parent entities lookup is required (#971)
On demand tiles generation requires that the request table contains all the
required serving entities (in order to know the entities for which to build the
tiles). This is not the case when additional parent entities need to be looked
up in the request table. This PR fixes the issue by adding the missing parent
entities lookup step before generating tiles.