[DEV-1289][DEV-1293][DEV-1294] documentation: update layout of docs (#977)
Make an assortment of changes based on Xavier's feedback, and also address some outstanding items.

Xavier's comments can be found here - https://featurebyte.atlassian.net/wiki/spaces/Docs/pages/87392438/Improvements+for+API+documentation

In addition, we also update the rendering of classes to only display the description and parameters, since displaying the properties and methods was noisy, and they're also replicated in other sections of the docs as needed. We also remove the links for classes for now since they don't really work since we don't generate docs for all of them right now. We can add them back in a more selective manner in a follow-up.