doc_layout: add missing titles in
We changed the way we obtained the title text in a previous commit
as we had to remove the `featurebyte.` prefix in the left nav bar
to improve the UX. This introduced a bug which caused some of the
titles to disappear when we had an explicit override.

This fixes that bug. Will look to add some test coverage into the
docs builder soon since this is stabilizing a bit more.