[DEV-1314][DEV-1322][DEV-1323] docs: fix missing create_*_table and update get_view docstring (#1001)
This PR makes a couple of changes in the docs.

We are currently missing the create_*_table methods in the documentation repo. This is due to a different setting with how docs are generated. In this repo, we pass in the FB_GENERATE_FULL_DOCS tag which will iterate through all classes and modules in featurebyte and try to generate some docs. However, this tag isn't passed in in the documentation repo, and we should try to avoid it as it might make building docs take a longer time. As such, we update the Makefile to be consistent with the documentation repo.

We then add the FBAutoDoc tag onto the SourceTable class to make sure that the class gets picked up for generating documentation.

Separately, we also:
- make some small docstring fixes in get_view per Xavier's feedback,
- add the FeatureGroup link in to make it consistent with the FeatureList item,
- add documentation for classes (taken from the Glossary - https://featurebyte.atlassian.net/wiki/spaces/Docs/pages/80412724/Glossary)