doc_layout: use featurestore methods for list and get
There's only one feature store per catalog, so it feels a bit
redundant to get a catalog, and then enumerate the feature store
from there. As such, we expose the FeatureStore#list and
FeatureStore#get as the main accessors for interacting with a
feature store, vs the catalog accessors, in the documentation.