[DEV-1341][DEV-1342][DEV-1343] docs: more minor updates (#1005)
A few more changes!

doc_layout: remove .cd, .str and .dt headers

These were previously placeholders before we had all the individual
methods in the docstrings. Now that they're in, we can remove
these headers.


doc_layout: use featurestore methods for list and get

There's only one feature store per catalog, so it feels a bit
redundant to get a catalog, and then enumerate the feature store
from there. As such, we expose the FeatureStore#list and
FeatureStore#get as the main accessors for interacting with a
feature store, vs the catalog accessors, in the documentation.


update styling to be same as documentation repo

Make the styling consistent with the documentation repo.