[DEV-1346] documentation_layout: simplify declaration of docs (#1031)
There was a bit of redundancy in the way we declared the DocLayoutItem. This was previously used when we had some discrepancies with the menu header, and the API path. Now that they're all the same, we can simplify the declaration here.

If we need to de-couple the two again, we can easily add back an optional parameter, but will remove it entirely for now to keep it simple.