[DEV-1367] Remove usage of AsyncMongoMockClient in integration tests (#1028)
This removes usage of AsyncMongoMockClient in integration tests, which is
causing the "RecursionError: maximum recursion depth exceeded while calling a
Python object" error in some cases.

Since we always start an actual test mongo server when running integration
tests, we can update the persistent fixture to use that instead. One dev facing
change is that in order to run the integration tests locally, we need to run

make test-setup

to start the test mongo server first. I think this is already the standard
workflow anyway, though previously that was not required if only running a
subset of the integration tests.