[DEV-1347] Hide current flag column from SCDView (#1025)
This hides the current flag column from SCDView in order to prevent users from
creating transforms or features using that column. To support this, filtering
of SCD table using the current flag column is pushed down to the lowest level
in sql generation, right when selecting columns from the SCD table.