Update TileBasedAggregationSpec to use construct_agg_result_name (#1034)
This updates TileBasedAggregationSpec to use the newly added helper function
construct_agg_result_name() to name internally generated columns for
consistency with other aggregation functions. These internal column names are
generated and used on the fly so there is no backward compatibility issues.
Most of the changes are updates to fixtures.