[DEV-1245] Support relational operations between Series and a fixed timestamp (#1039)
This adds support for relational operations between Series and a fixed
timestamp value.

This updates the value parameter in SingleValueNodeParameters to have a
stronger type (previously it was Any) and adds handling for non-native scalar
value types. This should also fix an issue where cleaning parameters are
incorrectly truncated as integers.