[DEV-1391] add ability to dump docs into a CSV (#1050)
This adds a configuration into mkdocs to dump all the documentation into a CSV.

This makes it easier to review in its entirety.

In the future, I'll de-couple this extraction from the mkdocs flow, and have it as a separate script instead. However, that requires a bit more refactoring which I'll do in the future.