docs: add warehouse details classes, and ability to point to markdown files in docs repo (#1092)
This PR makes a few changes

- small change of adding a few more classes into the doc layout
- bigger change of adding the ability for the docs generation code to point to static markdown 
files in the docs repo. They are pointed in two ways: (1) via an env variable. This will be used 
mainly for local development where you explicitly set a docs repo file path so that you can 
view the docs when running locally in the featurebyte repo. (2) via a relative file path. This 
will be used in the documentation repo when we try to build the docs. This is untested, but 
will hopefully work when I bring this change into that repo later.
- adds a unit test to verify that items exposed via featurebyte.__init__ are present in docs. 
It is likely that we want to expose anything that is automatically imported in the docs, so this 
should be useful to futureproof additions to the __init__ file.