[DEV-1486] Execute query with longer timeout for potentially long running queries (#1112)
This adds a method execute_query_long_running() which is the same as
execute_query() but with a longer default timeout. This is intended to be used
when executing queries that can be potentially long running, such as historical
requests. The default timeout for get_async_query_stream() is also updated to
be longer since that is used by historical requests.