update more docstrings for pydantic fields (#1129)
This PR updates more docstrings, this time focusing more on pydantic fields.

We create a docstring override dictionary for pydantic fields that are inherited. I chose to do 
this, vs just overriding the fields in child classes as it feels a bit more error prone for pydantic 
fields due to the way docstrings are populated via the description field for these pydantic fields. 
Having to redeclare the Field variable might be error prone as we might miss other fields that 
are also declared in the constructor (eg. default, allow_mutation etc.).

As such, I chose to just move the docstring overrides out to a separate dictionary.