Rewrite range join query to use only exact join conditions (#1131)
This updates the range join query used by tile based features to use only exact
join conditions. Previously, the join was using a condition that involved OR
which impacted the query performance. We saw from the query profile that the
job could not be parallelised well when the entity column has low cardinality.