api/*: update default of include_id to True for methods
We got some feedback from users that it's sometimes unclear where
they should be getting the ID's of commonly used objects from.
This makes a small change to set include_id to True for some of
the more common list functions.

We still leave the parameter there in case they want to exclude
it.