workflow: add workflow to diff docs (#1158)
This PR adds a workflow to diff the built documentation. This should us catch any unintentional
changes to the docs, and also let developers know when they are updating a docstring that is actually 
being surfaced to the public. This is useful since not all docstrings are public facing, and this check 
will help developers know.

Currently, the check will fail if there is a diff. If the changes look expected, simply add a 
`Documentation Changes Confirmed` label to the PR to skip the check.

Will try to make the html diff prettier in the future task.