[DEV-1547] docs: update docstrings on class fields (#1175)
Update parameter descriptions.

This was a little trickier to do as it required a nested parsing of resources. For example, we 
would need to parse the ResourceDetails for each parameter in the class, so that we could 
obtain the docstrings for each of the parameters and populate them accordingly.

This was tricky as there was some recursion logic that was a little hard to disentangle right 
now. As such, I've opted for a statically defined set of class names that we will do this nested 
parsing for. There aren't too many right now, and I'll look to fix this more cleanly in the future. 
However, in the interest of time, I'm going with this hacky approach for now.

We also do a small cleanup to remove the * parameter from being displayed in the parameter 
section as it doesn't provide that much value.

Also adds some basic docs for the BatchFeatureTable and BatchRequestTable API objects in.