[DEV-1238][DEV-1547] add feature store id to catalog  (#1188) (#1230)
This PR updates the catalog to be able to store feature store IDs on the catalog object. This will allow us to associate a feature store with the catalog when we create the catalog.

This would in turn allow us to simplify some API calls (eg. get_data_source) to not require users to pass in the feature_store_name whenever they are making the request.

Note that we persist a list of feature_store_ids in mongo, instead of a single feature_store_id as there are plans in the future to allow a single catalog to be associated with multiple feature stores.