ci: update lint check to not run safety on every PR (#1248)
This commit updates the lint workflow to not run lint-safety
on every PR. This is because most of the failures that come from
this check are not caused by the PR itself, but by changes that
are outside of our control (eg. a new vulnerability is introduced).

The current state will result in situations in which a new
vulnerability introduced externally will cause us to not be able
to merge any PRs - which is a loss in developer productivity.

Instead, we move this check to before we deploy a new staging
environment. If a failure happens here, we won't update our
staging environment, and a failure message will get pushed to slack.
We can then triage accordingly.