[DEV-1719] Call F_TIMESTAMP_TO_INDEX directly in tile generation sql (#1324)
This refactors the tile generation sql to call F_TIMESTAMP_TO_INDEX directly to
generate the final tile indices. This removes the assumption that there is a
fixed tile start date in the tile sql, and the resulting sql is actually
simpler.

This will enable an optimisation in tile cache to reduce unnecessary
recalculation of tiles (to be added in a next PR). In that optimisation,
different entities can have different tile start dates.