Bind spark-thrift-test to different ports (#1388)
The spark-thrift-test container is used by integration tests. When it is 
running, it is not possible start the featurebyte spark container since they 
bind to the same ports. This binds the ports used by the test container to 
different ones so these containers can co-exist. This should make development a 
bit more convenient (no need to stop the featurebyte services before running 
task test-setup and vice versa).