[DEV-1816] Update online store tables using insert instead of merge (#1423)
This changes the way online store tables are updated by using INSERT operation
instead of MERGE INTO. This allows concurrent updates to the same table without
blocking. A new service is added to keep track of the current version for each
aggregation result. This information will be used when constructing the query
to retrieve results from the online store tables.