[chore][poc] api_object/list: refactor list to a list handler (#1427)
This PR is a proof-of-concept for a refactor to extract list logic out of the ApiObject class, and into a separate handler.

Specifically, we also

move @staticmethods out of the class into api_object_util: benefit of this is that it's clearer that there are no dependencies on the class, and that these functions do not depend on any state. This makes it clearer when we want to reuse them.
encapsulate all the list function dependencies (eg. route, schema, fields, and foreign_key mappings) into the new handler object: benefit here is that it's clearer to a new reader what class variables are actually used to configure the list operation
encapsulate list functionality into the list_handler object: benefit is that it's easier to unit test this functionality without having to try to create an ApiObject itself.
Downsides are:

code becomes a bit more verbose (eg. new CatalogListHandler)