[DEV-1788] api/target: add basic API object for Target (#1426)
This PR adds the basic API object for working with a Target. This allows users to create a Target 
via a create function in the form of

fraud_flag_7d = Target.create(
  name='Fraudulent CC Transaction after 7 days',
  entities='CC Transaction',
  horizon="7d",
  blind_spot="1d",
)
This will also save the Target.

Will add more functionality (eg. series like functionality) in the future once we get the forward 
aggregates working.