[DEV-1827] Reduce unnecessary backfill computation when deploying features (#1432)
This reduces unnecessary backfill computation when deploying features. More
specifically:

1. When there is already an active tile job for an aggregation_id, backfilling
of historical tiles are not necessary because the tiles are kept up to date by
the tile job.

2. When populating online store tables, populate the results for a specific
aggregation result name, rather than for all the aggregation result names
belonging to the same aggregation_id. The latter makes sense within a scheduled
feature job but is entirely redundant when deploying a new feature.