[chore] app_container: refactor app_container to simplify (#1450)
With the introduction of the LazyAppContainer, and the smarter way of building our dependencies, we can actually simplify the API for the app_container, as the registration of dependencies is no longer order dependent (i.e. the order in which the registration happens in registry.py can be out of order, whereas previously it couldn't be).

As such, we can also simplify the registration API on the config to just 2 methods - register_service, and register_class. We still keep a special register_service since our services consistently take in a few parameters like user, persistent and catalog_id.

We also add a new validator to check for cycles in the dependencies when the app_container_config is built.