[chore] app_container_config: simplify interface further (#1484)
This PR updates the app container further, to simplify the interface in registering services to the config, to just a single method of register_class. We achieve this by pre-populating the instance_map with the user, persistent and catalog_id so that services registered via register_service, can now also just look up the user, persistent and catalog_id during construction like any other dependency.

We also make a small change to how names are parsed. Previously, a class with multiple upper case letters in a row will be parsed as such SCDTable -> s_c_d_table. This caused some errors when we tried to look up the dependencies, as we normally name our variables in the format scd_table instead. This change should alleviate the need for name_override's in this scenario.