[DEV-1824] target: add preview, and integration test (#1477)
This PR makes a few additions to the Target API.

adds a preview endpoint as a way to materialize some values of Target
adds an end-to-end integration test
have Target extend the Series class. Note that we'll still need to make some changes here to only allow Targets to interact with other Targets. Will also add more tests for this in a follow-up. Mostly added it here to unblock some of the code for the end-to-end test.