[DEV-1936] Remove TILE_FEATURE_MAPPING table by tracking aggregation_ids in feature collection (#1513)
This removes the TILE_FEATURE_MAPPING table in the data warehouse which was
used to determine whether a tile job can be stopped when a feature is online
disabled. The same information - list of aggregation_ids for each feature - is
now tracked in the feature collection and is available via the existing
FeatureService.