[chore] rm base service as it isn't needed (#1523)
There isn't really a need for the BaseService class as most of the services that inherit from it don't require a user, persistent, or catalog ID. The document services are mostly the ones that need references to those parameters. As such, we can remove them here, and inject them directly into the services that need them.